+++ deploy/terraform/modules/vpc/variables.tf
@@ -1,0 +1,4 @@
+variable "cidr_block" {
+  type        = string
+  description = "CIDR block for the isolated lab VPC."
+}